Creacion de carpetas con img
Se agregaron las imagenes para el README, se termino por comentar algunas partes del codigo y finalmente se cambio el tipo de dato del dinero de int a double.

+++ src/modelo/accionesJDBC.java
@@ -11,6 +11,14 @@ public class accionesJDBC extends conector{
     private ResultSet rs;
     private PreparedStatement ps;
     
+    
+    /**
+     * Metodo empleado para ingresar un nuevo usuario en la base de datos.
+     * @param persona
+     * @return Regresa true si se guardo el usuario, regresa false si
+     * no se pudo ingresar al nuevo usuario.
+     */
+    
     public boolean ingresarNuevoUsuario(usuarios persona){
     
         Connection conexion = getConnection();
@@ -38,6 +46,13 @@ public class accionesJDBC extends conector{
         }
     }
     
+    /**
+     * Metodo para encontrar usuarios en la base de datos
+     * @param persona
+     * @return Regresa true si se encontro al usuario o false si en caso contrario
+     * no se encontro el usuario en la base de datos.
+     */
+    
     public boolean encontrarUsuario(usuarios persona){
         Connection conexion = getConnection();
         try{
@@ -57,12 +72,21 @@ public class accionesJDBC extends conector{
         }
     }
     
-    public boolean consignar(usuarios persona, int cantidad){
+    /**
+     * Metodo para consignar a nuestra cuenta. 
+     * Introducimos el parametro persona para buscar al usuario y entroducimos 
+     * la cantidad a ignresar a nuestra cuenta.
+     * @param persona
+     * @param cantidad
+     * @return Regresa true si la operacion fue exitosa, false en caso contrario.
+     */
+    
+    public boolean consignar(usuarios persona, double cantidad){
         cantidad += persona.getDinero();
         Connection conexion = getConnection();
         try{
             ps = conexion.prepareStatement("update cowBank set dinero=? where nombre=?");
-            ps.setInt(1,cantidad);
+            ps.setDouble(1,cantidad);
             ps.setString(2, persona.nombre);
             int comprobacion = ps.executeUpdate();
             if(comprobacion > 0){
@@ -82,6 +106,12 @@ public class accionesJDBC extends conector{
         }
     }
     
+    /**
+     * Actualiza un usuario. Su uso es esencialmente para actualizar la cantidad de dinero de la cuenta
+     * @param persona
+     * @return Regresa true si la operacion fue exitosa, false en caso contrario.
+     */
+    
     public boolean actualizarUsuario(usuarios persona){
         Connection conexion = getConnection();
         try{
@@ -100,6 +130,13 @@ public class accionesJDBC extends conector{
             return false;
         }
     }
+    
+    /**
+     * Metodo que actualiza la cantidad de dinero del usuario, esencialmente pensada para retirar dinero.
+     * @param persona
+     * @param cantidad
+     * @return Regresa true si la operacion fue exitosa, false en caso contrario.
+     */
     
     public boolean retirar(usuarios persona, int cantidad){
